forEach loops active classes
Added for each loops to go through each panel. I also added an event listener to respond to the user clicking on an image and changing their class to active so that they adopt the flex of 5. I also added a removeActiveClasses function so that when a new picture is clicked, the previous one is no longer active so that the page highlights one card at a time